+++ es5/lists/getVersions.js
@@ -1,29 +1,18 @@
 import _regeneratorRuntime from "@babel/runtime-corejs3/regenerator";
-import _mapInstanceProperty from "@babel/runtime-corejs3/core-js-stable/instance/map";
-import _Array$isArray from "@babel/runtime-corejs3/core-js-stable/array/is-array";
 import _asyncToGenerator from "@babel/runtime-corejs3/helpers/esm/asyncToGenerator";
-import info from './info.js';
 import ajax from '../utils/ajax.js';
 /**
   @name $SP().list.getVersions
   @function
-  @description When versionning is activated on a list, you can use this function to get the different version label/number for a list item
+  @description When versionning is activated on a list, you can use this function to get the different versions of a list item
 
   @param {Number} ID The item ID
-  @return {Promise} resolve(arrayOflistOfVersions)
+  @return {Promise} resolve(arrayOfVersions)
 
   @example
-  $SP().list("My List").getVersions({
-    ID:1
-  }).then(function(versions) {
+  $SP().list("My List").getVersions(1234).then(function(versions) {
     versions.forEach(function(version) {
       console.log(version);
-      // returns:
-      //  - CheckInComment
-      //  - Created
-      //  - VersionID
-      //  - IsCurrentVersion (boolean)
-      //  - VersionLabel (e.g. "1.0", "2.0", …)
     })
   });
 */
@@ -36,13 +25,12 @@ function _getVersions() {
   _getVersions = _asyncToGenerator(
   /*#__PURE__*/
   _regeneratorRuntime.mark(function _callee(itemID) {
-    var infos, rootFolder;
-    return _regeneratorRuntime.wrap(function _callee$(_context2) {
+    return _regeneratorRuntime.wrap(function _callee$(_context) {
       while (1) {
-        switch (_context2.prev = _context2.next) {
+        switch (_context.prev = _context.next) {
           case 0:
             if (this.listID) {
-              _context2.next = 2;
+              _context.next = 2;
               break;
             }
 
@@ -50,7 +38,7 @@ function _getVersions() {
 
           case 2:
             if (this.url) {
-              _context2.next = 4;
+              _context.next = 4;
               break;
             }
 
@@ -58,48 +46,22 @@ function _getVersions() {
 
           case 4:
             if (itemID) {
-              _context2.next = 6;
+              _context.next = 6;
               break;
             }
 
             throw "[SharepointPlus 'getVersions'] the item ID is required.";
 
           case 6:
-            _context2.next = 8;
-            return info.call(this);
-
-          case 8:
-            infos = _context2.sent;
-            rootFolder = infos._List.RootFolder; // if no versionning
-
-            if (!(infos._List.EnableVersioning !== "True")) {
-              _context2.next = 12;
-              break;
-            }
-
-            return _context2.abrupt("return", []);
-
-          case 12:
-            return _context2.abrupt("return", ajax.call(this, {
-              url: this.url + "/_api/web/GetFileByServerRelativeUrl('" + encodeURIComponent(rootFolder + "/" + itemID + "_.000") + "')/Versions"
+            return _context.abrupt("return", ajax.call(this, {
+              url: this.url + "/_api/lists/getbytitle('" + this.listID + "')/Items(" + itemID + ")/Versions"
             }).then(function (res) {
-              var _context;
-
-              if (!res || !res.d || !_Array$isArray(res.d.results)) return [];
-              return _mapInstanceProperty(_context = res.d.results).call(_context, function (item) {
-                return {
-                  CheckInComment: item.CheckInComment,
-                  Created: item.Created,
-                  VersionID: item.ID,
-                  IsCurrentVersion: item.IsCurrentVersion,
-                  VersionLabel: item.VersionLabel
-                };
-              });
+              return (res.d ? res.d.results : res.value) || [];
             }));
 
-          case 13:
+          case 7:
           case "end":
-            return _context2.stop();
+            return _context.stop();
         }
       }
     }, _callee, this);