+++ es5/lists/removeAttachment.js
@@ -17,7 +17,7 @@ import restoreVersion from './restoreVersion.js';
   @example
   $SP().list("My List").removeAttachment({
     ID:1,
-    filename:"https://mysite.share.point.com/Toolbox/Lists/Tasks/Attachments/2305/image1.png"
+    fileURL:"https://mysite.share.point.com/Toolbox/Lists/Tasks/Attachments/2305/image1.png"
   })
 */