+++ es5/main.js
@@ -27,6 +27,7 @@ global._SP_PLUGINS = {};
 global._SP_MODALDIALOG_LOADED = false;
 global._SP_MAXWHERE_ONLOOKUP = 30;
 global._SP_ISBROWSER = new Function("try {return this===window;}catch(e){ return false;}")();
+global._SP_ISSPO = {};
 global._SP_JSON_ACCEPT = "verbose"; // other options are "minimalmetadata" and "nometadata"
 
 var SharepointPlus =
@@ -60,7 +61,7 @@ function () {
   _createClass(SharepointPlus, [{
     key: "getVersion",
     value: function getVersion() {
-      return "6.1.5";
+      return "6.2.0";
     }
     /**
       @name $SP().auth