+++ src/people/getManager.js
@@ -8,13 +8,14 @@ import people from './people.js'
   @category people
   @description Return the manager for the provided user, as a People string
 
-  @param {String} [username] With or without the domain, and you can also use an email address, and if you leave it empty it's the current user by default
+  @param {String} [username] Username with the domain, and if you leave it empty it's the current user by default
   @param {Object} [setup] Options (see below)
     @param {String} [setup.url='current website'] The website url
+    @param {Function} [setup.modify] Permits to modify the manager's username returned by the service
   @return {Function} resolve(manager), reject(error)
 
   @example
-  $SP().getManager("john_doe",{url:"http://my.si.te/subdir/"})
+  $SP().getManager("domain\\john_doe",{url:"http://my.si.te/subdir/"})
   .then(function(manager) {
     console.log(manager); // 42;#Smith,, Jane,#i:0#.w|domain\Jane_Smith,#Jane_Smith@Domain.com,#Jane_Smith@Domain.com,#Smith,, Jane
     manager = $SP().getPeopleLookup(manager);
@@ -23,6 +24,12 @@ import people from './people.js'
   .catch(function(err) {
     console.log("Err => ",err)
   });
+
+  $SP().getManager("domain\\john_doe",{
+    modify:function(managerUserName) {
+      return (managerUserName.startsWith('i:0') ? managerUserName : "i:0#.w|" + managerUserName);
+    }
+  })
 */
 export default async function getManager(username, setup) {
   try {
@@ -33,10 +40,10 @@ export default async function getManager(username, setup) {
     if (!setup.url) {
       setup.url = await getURL.call(this);
     }
+    setup.modify = setup.modify || function(val) { return val };
 
     let pres = await people.call(this, username, setup);
-    let managerUserName = pres.Manager;
-    if (!managerUserName.startsWith('i:0')) managerUserName = "i:0#.w|" + managerUserName;
+    let managerUserName = setup.modify(pres.Manager);
     let res = await getUserInfo.call(this, managerUserName, setup);
 
     // "42;#Doe,, John,#i:0#.w|domain\John_Doe,#John_Doe@Domain.com,#John_Doe@Domain.com,#Doe,, John