+++ src/people/isMember.js
@@ -33,7 +33,7 @@ export default async function isMember(setup) {
     setup.group = setup.group.toLowerCase();
     let members=[];
     // first check with usergroups()
-    let groups = await usergroups.call(this, setup.user, {cache:setup.cache})
+    let groups = await usergroups.call(this, setup.user, {cache:setup.cache, url:setup.url})
     for (let i=groups.length; i--;) {
       if (groups[i].toLowerCase() === setup.group) {
         return Promise.resolve(true);
@@ -42,11 +42,11 @@ export default async function isMember(setup) {
 
     // if we're there then it means we need to keep investigating
     // look at the members of the group
-    let m = await groupMembers.call(this, setup.group, {cache:setup.cache});
+    let m = await groupMembers.call(this, setup.group, {cache:setup.cache, url:setup.url});
     for (let i=m.length; i--;) members.push(m[i].Name.toLowerCase());
 
     // and search if our user is part of the members (like a distribution list)
-    let distrib = await distributionLists.call(this, setup.user, {cache:setup.cache});
+    let distrib = await distributionLists.call(this, setup.user, {cache:setup.cache, url:setup.url});
 
     for (let i=distrib.length; i--;) {
       if (members.indexOf(distrib[i].DisplayName.toLowerCase()) > -1) {