+++ src/utils/hasREST.js
@@ -1,12 +1,11 @@
-import ajax from './ajax.js'
 
 /**
   @name $SP().hasREST
   @function
   @category utils
-  @description Verify if the website supports REST API (Sharepoint 2013 and later)
+  @description In the earlier version of SharePointPlus, this function was used to check if the REST API was available – these days I assume everyone is now using at least SharePoint 2013, so this function always returns TRUE – if you don't have REST API you can still define _SP_CACHE_HASREST["url to check"]=false
   @param {Object} settings
-    @param {String} [settings.url=current] To check another URL (or if you need on a Node server)
+    @param {String} [settings.url=current] To check another URL
   @return {Promise} A resolved Promise that gives TRUE or FALSE
 */
 export default function hasREST(settings) {
@@ -16,28 +15,5 @@ export default function hasREST(settings) {
   if (typeof global._SP_CACHE_HASREST[url] === "boolean") {
     return Promise.resolve(global._SP_CACHE_HASREST[url]);
   }
-  let hasREST, needAjax=(settings.url || !global._SP_ISBROWSER || typeof SP === "undefined" ? true : false);
-  if (!needAjax) {
-    if (typeof SP !== "undefined" && SP.ClientSchemaVersions) { // eslint-disable-line
-      // cache
-      hasREST=(parseInt(SP.ClientSchemaVersions.currentVersion)>14); // eslint-disable-line
-      global._SP_CACHE_HASREST[url]=hasREST;
-      return Promise.resolve(hasREST);
-    }
-    else needAjax=true;
-  }
-  if (needAjax) {
-    return ajax.call(this, {url:url + "/_api/web/Url"})
-    .then(function() {
-      global._SP_CACHE_HASREST[url]=true;
-      return Promise.resolve(true)
-    })
-    .catch(function() {
-      global._SP_CACHE_HASREST[url]=false;
-      return Promise.resolve(false)
-    })
-  } else {
-    global._SP_CACHE_HASREST[url]=false;
-    return Promise.resolve(false);
-  }
+  return Promise.resolve(true);
 }